+++ models/classInfo.go
@@ -2,8 +2,8 @@ package models
 
 type ClassInfo struct {
 	Name          string                `json:"name" firestore:"name"`
-	Level         int32                 `json:"level" firestore:"level"`
-	Exp           int32                 `json:"exp" firestore:"exp"`
+	Level         int64                 `json:"level" firestore:"level"`
+	Exp           int64                 `json:"exp" firestore:"exp"`
 	CurrentWeapon string                `json:"currentWeapon" firestore:"currentWeapon"`
 	Equipment     Equipment             `json:"equipment" firestore:"equipment"`
 	BossBonuses   map[string]*BossBonus `json:"bossBonuses" firestore:"bossBonuses"`
