+++ resolvers/classResolver.go
@@ -15,12 +15,12 @@ func (c *classResolver) Name(_ context.Context) string {
 	return c.classInfo.Name
 }
 
-func (c *classResolver) Level(_ context.Context) int32 {
-	return c.classInfo.Level
+func (c *classResolver) Level(_ context.Context) float64 {
+	return float64(c.classInfo.Level)
 }
 
-func (c *classResolver) Exp(_ context.Context) int32 {
-	return c.classInfo.Exp
+func (c *classResolver) Exp(_ context.Context) float64 {
+	return float64(c.classInfo.Exp)
 }
 
 func (c *classResolver) Equipment(_ context.Context) *equipmentResolver {
